refactor: split god modules into focused files
Split 3 large files (>500 lines) into smaller, single-responsibility modules:

rpc/logs.go (833 -> 398 lines):
- logs_store.go: store interfaces + LogsStoreFuncs adapter (123 lines)
- logs_ring.go: localLogRing buffer implementation (85 lines)
- logs_proto.go: proto conversion helpers (170 lines)

commands/reqs.go (885 -> 260 lines):
- reqs_checker.go: PrerequisiteChecker + command validation (303 lines)
- reqs_version.go: version parsing/comparison utilities (126 lines)
- reqs_fix.go: fix runner flow (213 lines)

commands/run.go (769 -> 328 lines):
- run_orchestration.go: service execution/monitoring lifecycle (347 lines)
- run_hooks.go: hook execution/conversion (112 lines)

All files now under 500 lines. No behavioral changes.
