+++ cli/src/cmd/app/commands/reqs.go
@@ -1,18 +1,11 @@
 package commands
 
 import (
-	"context"
 	"fmt"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"regexp"
-	"runtime"
 	"strings"
 
-	"github.com/jongio/azd-app/cli/src/internal/cache"
 	"github.com/jongio/azd-core/cliout"
-	"github.com/jongio/azd-core/pathutil"
 
 	"github.com/spf13/cobra"
 )
@@ -175,7 +168,7 @@ var toolRegistry = map[string]ToolConfig{
 	},
 	"az": {
 		Command: "az",
-		Args:    []string{"version", "--output", "tsv", "--query", "\"azure-cli\""},
+		Args:    []string{"version", "--output", "tsv", "--query", `"azure-cli"`},
 	},
 	"air": {
 		Command:       "air",
@@ -189,7 +182,7 @@ var toolRegistry = map[string]ToolConfig{
 	"java": {
 		Command:      "java",
 		Args:         []string{"-version"},
-		VersionField: 2, // "java version "17.0.1" ..." or "openjdk version "17.0.1" ..." -> take field 2
+		VersionField: 2, // "java version \"17.0.1\" ..." or "openjdk version \"17.0.1\" ..." -> take field 2
 	},
 	"mvn": {
 		Command:      "mvn",
@@ -208,86 +201,6 @@ var toolAliases = map[string]string{
 	"nodejs":                     "node",
 	"azure-cli":                  "az",
 	"azure-functions-core-tools": "func",
-}
-
-var allowedCustomPrereqCommands = map[string]struct{}{
-	"air":      {},
-	"aspire":   {},
-	"az":       {},
-	"azd":      {},
-	"cargo":    {},
-	toolDocker: {},
-	"dotnet":   {},
-	"func":     {},
-	"git":      {},
-	"go":       {},
-	"gradle":   {},
-	"java":     {},
-	"mvn":      {},
-	"node":     {},
-	"npm":      {},
-	"pip":      {},
-	"pipenv":   {},
-	"pnpm":     {},
-	"poetry":   {},
-	"python":   {},
-	"uv":       {},
-	"yarn":     {},
-}
-
-func normalizeAllowedCommandName(command string) string {
-	trimmed := strings.TrimSpace(command)
-	if trimmed == "" || strings.ContainsAny(trimmed, `\\/`) {
-		return ""
-	}
-
-	name := strings.ToLower(filepath.Base(trimmed))
-	return strings.TrimSuffix(name, ".exe")
-}
-
-func isAllowedCustomPrereqCommand(command string) bool {
-	name := normalizeAllowedCommandName(command)
-	if name == "" {
-		return false
-	}
-	_, allowed := allowedCustomPrereqCommands[name]
-	return allowed
-}
-
-func validateCustomPrereqCommand(fieldName, command string) error {
-	if strings.TrimSpace(command) == "" {
-		return nil
-	}
-	if !isAllowedCustomPrereqCommand(command) {
-		return fmt.Errorf("invalid %s %q: only allowlisted tool commands are permitted", fieldName, command)
-	}
-	return nil
-}
-
-// installURLRegistry maps tool names to their installation page URLs.
-var installURLRegistry = map[string]string{
-	"node":     "https://nodejs.org/",
-	"npm":      "https://nodejs.org/",
-	"pnpm":     "https://pnpm.io/installation",
-	"yarn":     "https://yarnpkg.com/getting-started/install",
-	"python":   "https://www.python.org/downloads/",
-	"pip":      "https://www.python.org/downloads/",
-	"poetry":   "https://python-poetry.org/docs/#installation",
-	"uv":       "https://docs.astral.sh/uv/getting-started/installation/",
-	"pipenv":   "https://pipenv.pypa.io/en/latest/installation.html",
-	"dotnet":   "https://dotnet.microsoft.com/download",
-	"aspire":   "https://learn.microsoft.com/dotnet/aspire/fundamentals/setup-tooling",
-	toolDocker: "https://www.docker.com/products/docker-desktop",
-	"git":      "https://git-scm.com/downloads",
-	"go":       "https://go.dev/dl/",
-	"azd":      "https://aka.ms/install-azd",
-	"az":       "https://aka.ms/installazurecli",
-	"air":      "https://github.com/air-verse/air#installation",
-	"func":     "https://learn.microsoft.com/azure/azure-functions/functions-run-local#install-the-azure-functions-core-tools",
-	"java":     "https://adoptium.net/",
-	"mvn":      "https://maven.apache.org/install.html",
-	"gradle":   "https://gradle.org/install/",
-	"gh":       "https://cli.github.com/",
 }
 
 // NewReqsCommand creates the reqs command.
@@ -371,632 +284,4 @@ Use --no-cache to force a fresh check and bypass cached results.`,
 func runReqs() error {
 	// Use orchestrator to execute reqs check with caching support
 	return executeReqs()
-}
-
-// PrerequisiteChecker handles checking of prerequisites.
-type PrerequisiteChecker struct {
-	registry map[string]ToolConfig
-	aliases  map[string]string
-}
-
-// NewPrerequisiteChecker creates a new prerequisite checker.
-func NewPrerequisiteChecker() *PrerequisiteChecker {
-	return &PrerequisiteChecker{
-		registry: toolRegistry,
-		aliases:  toolAliases,
-	}
-}
-
-// Check checks a prerequisite and returns structured result.
-func (pc *PrerequisiteChecker) Check(prereq Prerequisite) ReqResult {
-	// Resolve install URL (custom overrides built-in)
-	installURL := pc.getInstallURL(prereq)
-
-	result := ReqResult{
-		Name:       prereq.Name,
-		Required:   prereq.MinVersion,
-		Satisfied:  false,
-		InstallURL: installURL,
-	}
-
-	if err := validateCustomPrereqCommand("command", prereq.Command); err != nil {
-		result.Message = err.Error()
-		return result
-	}
-	if prereq.CheckRunning {
-		if err := validateCustomPrereqCommand("runningCheckCommand", prereq.RunningCheckCommand); err != nil {
-			result.Message = err.Error()
-			return result
-		}
-	}
-
-	installed, version, isPodman := pc.getInstalledVersion(prereq)
-	result.Installed = installed
-	result.Version = version
-	result.IsPodman = isPodman
-
-	if !installed {
-		result.Message = "Not installed"
-		if !cliout.IsJSON() {
-			cliout.ItemError("%s: NOT INSTALLED (required: %s)", prereq.Name, prereq.MinVersion)
-			if installURL != "" {
-				cliout.Item("   Install: %s", installURL)
-			}
-		}
-		return result
-	}
-
-	// When Podman is aliased to Docker, skip version comparison since version schemes differ.
-	// Podman uses its own versioning (e.g., 5.7.0) which is not comparable to Docker versions (e.g., 20.10.0).
-	if isPodman && prereq.Name == toolDocker {
-		result.Message = "Podman detected (version check skipped)"
-		if !cliout.IsJSON() {
-			cliout.ItemSuccess("%s: %s via Podman (version check skipped)", prereq.Name, version)
-		}
-		// Continue to check if running if needed, otherwise mark satisfied
-		if !prereq.CheckRunning {
-			result.Satisfied = true
-			return result
-		}
-	} else if version == "" {
-		result.Message = "Version unknown"
-		if !cliout.IsJSON() {
-			cliout.ItemWarning("%s: INSTALLED (version unknown, required: %s)", prereq.Name, prereq.MinVersion)
-		}
-		// Continue to check if it's running if needed
-	} else {
-		versionOk := compareVersions(version, prereq.MinVersion)
-		if !versionOk {
-			result.Message = fmt.Sprintf("Version %s does not meet minimum %s", version, prereq.MinVersion)
-			if !cliout.IsJSON() {
-				cliout.ItemError("%s: %s (required: %s)", prereq.Name, version, prereq.MinVersion)
-				if installURL != "" {
-					cliout.Item("   Install: %s", installURL)
-				}
-			}
-			return result
-		}
-		if !cliout.IsJSON() {
-			cliout.ItemSuccess("%s: %s (required: %s)", prereq.Name, version, prereq.MinVersion)
-		}
-	}
-
-	// Check if the tool is running (if configured)
-	if prereq.CheckRunning {
-		result.CheckedRun = true
-		isRunning := pc.checkIsRunning(prereq)
-		result.Running = isRunning
-		if !isRunning {
-			result.Message = "Not running"
-			if !cliout.IsJSON() {
-				cliout.Item("- %s✗%s NOT RUNNING", cliout.Red, cliout.Reset)
-			}
-			return result
-		}
-		result.Satisfied = true
-		result.Message = "Running"
-		if !cliout.IsJSON() {
-			cliout.Item("- %s✓%s RUNNING", cliout.Green, cliout.Reset)
-		}
-		return result
-	}
-
-	if version != "" {
-		result.Satisfied = true
-		result.Message = "Satisfied"
-	}
-	return result
-}
-
-// getInstallURL returns the install URL for a prerequisite.
-// Custom InstallURL in prerequisite takes precedence over built-in registry.
-func (pc *PrerequisiteChecker) getInstallURL(prereq Prerequisite) string {
-	// Custom URL takes precedence
-	if prereq.InstallURL != "" {
-		return prereq.InstallURL
-	}
-
-	// Resolve aliases to canonical name
-	tool := prereq.Name
-	if canonical, isAlias := pc.aliases[tool]; isAlias {
-		tool = canonical
-	}
-
-	// Look up in registry
-	if url, found := installURLRegistry[tool]; found {
-		return url
-	}
-
-	return ""
-}
-
-// getInstalledVersion gets the installed version of a prerequisite.
-// Returns isPodman=true when Podman is detected aliased to Docker.
-func (pc *PrerequisiteChecker) getInstalledVersion(prereq Prerequisite) (installed bool, version string, isPodman bool) {
-	config := pc.getToolConfig(prereq)
-
-	// #nosec G204 -- Command and args come from toolRegistry or validated azure.yaml prerequisite configuration
-	cmd := exec.CommandContext(context.Background(), config.Command, config.Args...)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return false, "", false
-	}
-
-	outputStr := strings.TrimSpace(string(output))
-
-	// Detect Podman aliased to Docker
-	isPodman = strings.Contains(outputStr, "Podman Engine")
-
-	version = extractVersion(config, outputStr)
-
-	return true, version, isPodman
-}
-
-// getToolConfig gets the tool configuration for a prerequisite.
-func (pc *PrerequisiteChecker) getToolConfig(prereq Prerequisite) ToolConfig {
-	// Check if custom configuration is provided in prerequisite
-	if prereq.Command != "" {
-		return ToolConfig{
-			Command:       normalizeAllowedCommandName(prereq.Command),
-			Args:          prereq.Args,
-			VersionPrefix: prereq.VersionPrefix,
-			VersionField:  prereq.VersionField,
-		}
-	}
-
-	// Use registry-based configuration
-	tool := prereq.Name
-
-	// Resolve aliases to canonical name
-	if canonical, isAlias := pc.aliases[tool]; isAlias {
-		tool = canonical
-	}
-
-	// Look up tool configuration
-	if config, found := pc.registry[tool]; found {
-		return config
-	}
-
-	// Fallback: try generic --version with tool ID as command
-	return ToolConfig{
-		Command: prereq.Name,
-		Args:    []string{"--version"},
-	}
-}
-
-// checkIsRunning checks if a prerequisite tool is currently running.
-func (pc *PrerequisiteChecker) checkIsRunning(prereq Prerequisite) bool {
-	// If no custom running check is configured, use defaults based on tool ID
-	command := prereq.RunningCheckCommand
-	args := prereq.RunningCheckArgs
-	expectedExitCode := 0
-	if prereq.RunningCheckExitCode != nil {
-		expectedExitCode = *prereq.RunningCheckExitCode
-	}
-
-	// Default checks for known tools
-	if command == "" {
-		switch prereq.Name {
-		case toolDocker:
-			command = toolDocker
-			args = []string{"ps"}
-		default:
-			// No default running check for this tool
-			// Return false to indicate check is not configured properly
-			// Users should provide RunningCheckCommand if checkRunning is true
-			return false
-		}
-	}
-
-	if prereq.RunningCheckCommand != "" {
-		command = normalizeAllowedCommandName(command)
-	}
-
-	// #nosec G204 -- Command and args are validated against an allowlist or come from the built-in Docker check
-	cmd := exec.CommandContext(context.Background(), command, args...)
-	output, err := cmd.CombinedOutput()
-
-	// Check exit code
-	exitCode := 0
-	if err != nil {
-		if exitErr, ok := err.(*exec.ExitError); ok {
-			exitCode = exitErr.ExitCode()
-		} else {
-			// Command failed to execute
-			return false
-		}
-	}
-
-	if exitCode != expectedExitCode {
-		return false
-	}
-
-	// If an expected substring is configured, check for it in the output
-	if prereq.RunningCheckExpected != "" {
-		outputStr := strings.TrimSpace(string(output))
-		return strings.Contains(outputStr, prereq.RunningCheckExpected)
-	}
-
-	return true
-}
-
-// Deprecated: Use PrerequisiteChecker.Check instead
-func checkPrerequisite(prereq Prerequisite) bool {
-	checker := NewPrerequisiteChecker()
-	result := checker.Check(prereq)
-	return result.Satisfied
-}
-
-// extractVersion extracts version from command cliout.
-func extractVersion(config ToolConfig, output string) string {
-	// Handle azd special case first (multi-line output)
-	if strings.Contains(output, "azd version") {
-		return extractAzdVersion(output)
-	}
-
-	// Handle Podman aliased to Docker (multi-line output with "Podman Engine")
-	if strings.Contains(output, "Podman Engine") {
-		return extractPodmanVersion(output)
-	}
-
-	// Extract specific field BEFORE stripping prefix (field extraction first)
-	if config.VersionField > 0 {
-		parts := strings.Fields(output)
-		if len(parts) > config.VersionField {
-			output = parts[config.VersionField]
-		}
-	}
-
-	// Strip prefix if configured (after field extraction)
-	if config.VersionPrefix != "" {
-		output = strings.TrimPrefix(output, config.VersionPrefix)
-	}
-
-	return extractFirstVersion(output)
-}
-
-// extractAzdVersion extracts version from azd multi-line cliout.
-func extractAzdVersion(output string) string {
-	lines := strings.Split(output, "\n")
-	for _, line := range lines {
-		if strings.Contains(line, "azd version") {
-			parts := strings.Fields(line)
-			for _, part := range parts {
-				if v := extractFirstVersion(part); v != "" && v != "version" {
-					return v
-				}
-			}
-		}
-	}
-	return ""
-}
-
-// extractPodmanVersion extracts version from Podman multi-line cliout.
-// Podman output format when aliased to docker:
-//
-//	Client:       Podman Engine
-//	Version:      5.7.0
-//	API Version:  5.7.0
-//	...
-func extractPodmanVersion(output string) string {
-	lines := strings.Split(output, "\n")
-	for _, line := range lines {
-		line = strings.TrimSpace(line)
-		if strings.HasPrefix(line, "Version:") {
-			// Extract version from "Version:      5.7.0"
-			parts := strings.Fields(line)
-			if len(parts) >= 2 {
-				return extractFirstVersion(parts[1])
-			}
-		}
-	}
-	return ""
-}
-
-// Compiled regex patterns for version extraction (package-level for performance)
-var (
-	semanticVersionRegex = regexp.MustCompile(`(\d+\.\d+\.\d+)`)
-	simpleVersionRegex   = regexp.MustCompile(`(\d+\.\d+)`)
-)
-
-// extractFirstVersion finds the first semantic version in a string.
-func extractFirstVersion(s string) string {
-	// Match semantic version pattern (e.g., 1.2.3, 20.0.0, etc.)
-	matches := semanticVersionRegex.FindStringSubmatch(s)
-	if len(matches) > 1 {
-		return matches[1]
-	}
-
-	// Try simpler pattern (e.g., 1.2)
-	matches = simpleVersionRegex.FindStringSubmatch(s)
-	if len(matches) > 1 {
-		return matches[1]
-	}
-
-	return ""
-}
-
-// compareVersions compares installed version against required version.
-// Returns true if installed >= required.
-// Missing version parts are treated as 0 (e.g., "1.2" is equivalent to "1.2.0").
-func compareVersions(installed, required string) bool {
-	installedParts := parseVersion(installed)
-	requiredParts := parseVersion(required)
-
-	// Get the maximum length to compare all parts
-	maxLen := len(requiredParts)
-	if len(installedParts) > maxLen {
-		maxLen = len(installedParts)
-	}
-
-	// Compare each part left to right, treating missing parts as 0
-	for i := 0; i < maxLen; i++ {
-		installedPart := 0
-		if i < len(installedParts) {
-			installedPart = installedParts[i]
-		}
-
-		requiredPart := 0
-		if i < len(requiredParts) {
-			requiredPart = requiredParts[i]
-		}
-
-		if installedPart > requiredPart {
-			return true
-		}
-		if installedPart < requiredPart {
-			return false
-		}
-		// Equal, continue to next part
-	}
-
-	return true // All parts equal
-}
-
-// parseVersion parses a version string into numeric parts.
-func parseVersion(version string) []int {
-	parts := strings.Split(version, ".")
-	result := make([]int, 0, len(parts))
-
-	for _, part := range parts {
-		var num int
-		_, _ = fmt.Sscanf(part, "%d", &num)
-		result = append(result, num)
-	}
-
-	return result
-}
-
-// runClearCache clears the reqs cache.
-func runClearCache() error {
-	cacheManager, err := cache.NewCacheManager()
-	if err != nil {
-		return fmt.Errorf("failed to initialize cache manager: %w", err)
-	}
-
-	if err := cacheManager.ClearCache(); err != nil {
-		return fmt.Errorf("failed to clear cache: %w", err)
-	}
-
-	if cliout.IsJSON() {
-		return cliout.PrintJSON(map[string]any{
-			"success": true,
-			"message": "Reqs cache cleared successfully",
-		})
-	}
-
-	cliout.Success("Reqs cache cleared successfully")
-	return nil
-}
-
-// FixResult represents the result of attempting to fix a requirement.
-type FixResult struct {
-	Name      string `json:"name"`
-	Fixed     bool   `json:"fixed"`
-	Found     bool   `json:"found"`
-	Path      string `json:"path,omitempty"`
-	Message   string `json:"message"`
-	Satisfied bool   `json:"satisfied"`
-}
-
-type reqsFixRunner struct {
-	checker *PrerequisiteChecker
-}
-
-func newReqsFixRunner() *reqsFixRunner {
-	return &reqsFixRunner{checker: NewPrerequisiteChecker()}
-}
-
-func (r *reqsFixRunner) parseReqs() (string, []Prerequisite, error) {
-	azureYamlPath, azureYaml, err := loadAzureYaml()
-	if err != nil {
-		return "", nil, err
-	}
-
-	reqs := append([]Prerequisite{}, azureYaml.Reqs...)
-	reqs = r.ensureDockerReq(azureYaml, reqs)
-	if len(reqs) == 0 {
-		return "", nil, fmt.Errorf("no reqs defined in azure.yaml - run 'azd app reqs --generate' to add them")
-	}
-
-	return azureYamlPath, reqs, nil
-}
-
-func (r *reqsFixRunner) ensureDockerReq(azureYaml *AzureYaml, reqs []Prerequisite) []Prerequisite {
-	if azureYaml.hasContainerServices() && !azureYaml.hasDockerReq() {
-		return append(reqs, Prerequisite{Name: "docker", MinVersion: "20.0.0", CheckRunning: true})
-	}
-	return reqs
-}
-
-func (r *reqsFixRunner) checkPrerequisite(prereq Prerequisite) FixResult {
-	fixResult := FixResult{Name: prereq.Name}
-	config := r.checker.getToolConfig(prereq)
-	toolCommand := config.Command
-
-	toolPath := pathutil.FindToolInPath(toolCommand)
-	if toolPath != "" {
-		fixResult.Found = true
-		fixResult.Path = toolPath
-		result := r.checker.Check(prereq)
-		if result.Satisfied {
-			fixResult.Fixed = true
-			fixResult.Satisfied = true
-			fixResult.Message = fmt.Sprintf("Found and verified: %s", toolPath)
-			if !cliout.IsJSON() {
-				cliout.ItemSuccess("Found: %s", toolPath)
-				cliout.ItemSuccess("Version verified successfully")
-			}
-			return fixResult
-		}
-		fixResult.Message = fmt.Sprintf("Found at %s but version check failed: %s", toolPath, result.Message)
-		if !cliout.IsJSON() {
-			cliout.ItemWarning("Found: %s", toolPath)
-			cliout.ItemWarning("Version check failed: %s", result.Message)
-		}
-		return fixResult
-	}
-
-	toolPath = pathutil.SearchToolInSystemPath(toolCommand)
-	if toolPath != "" {
-		fixResult.Found = true
-		fixResult.Path = toolPath
-		fixResult.Message = fmt.Sprintf("Found at %s but not in PATH - restart terminal may be needed", toolPath)
-		if !cliout.IsJSON() {
-			cliout.ItemWarning("Found: %s", toolPath)
-			cliout.ItemWarning("Tool is installed but not in current PATH")
-			cliout.Info("   %s Restart your terminal to update PATH", cliout.IconBulb)
-		}
-		return fixResult
-	}
-
-	suggestion := pathutil.GetInstallSuggestion(toolCommand)
-	fixResult.Message = fmt.Sprintf("Not found - %s", suggestion)
-	if !cliout.IsJSON() {
-		cliout.ItemError("Not found in system PATH")
-		cliout.Info("   %s %s", cliout.IconBulb, suggestion)
-	}
-	return fixResult
-}
-
-// runReqsFix attempts to fix PATH issues for missing tools.
-func runReqsFix() error {
-	cliout.CommandHeader("reqs --fix", "Fix PATH issues for missing tools")
-	if !cliout.IsJSON() {
-		cliout.Section(cliout.IconTool, "Attempting to fix requirement issues...")
-	}
-
-	runner := newReqsFixRunner()
-	azureYamlPath, reqs, err := runner.parseReqs()
-	if err != nil {
-		return err
-	}
-
-	var failedReqs []Prerequisite
-	for _, prereq := range reqs {
-		if result := runner.checker.Check(prereq); !result.Satisfied {
-			failedReqs = append(failedReqs, prereq)
-		}
-	}
-
-	if len(failedReqs) == 0 {
-		if cliout.IsJSON() {
-			return cliout.PrintJSON(map[string]any{"success": true, "message": "All requirements already satisfied"})
-		}
-		cliout.Success("All requirements already satisfied!")
-		return nil
-	}
-
-	if !cliout.IsJSON() {
-		cliout.Newline()
-		cliout.Step(cliout.IconRefresh, "Refreshing environment PATH...")
-	}
-	if _, err = pathutil.RefreshPATH(); err != nil {
-		if !cliout.IsJSON() {
-			cliout.Warning("Failed to refresh PATH: %v", err)
-		}
-	} else if !cliout.IsJSON() {
-		cliout.ItemSuccess("PATH refreshed successfully")
-	}
-
-	fixResults := make([]FixResult, 0, len(failedReqs))
-	fixedCount := 0
-	for _, prereq := range failedReqs {
-		if !cliout.IsJSON() {
-			cliout.Newline()
-			cliout.Step(cliout.IconSearch, "Searching for %s...", prereq.Name)
-		}
-
-		fixResult := runner.checkPrerequisite(prereq)
-		if fixResult.Fixed && fixResult.Satisfied {
-			fixedCount++
-		}
-		fixResults = append(fixResults, fixResult)
-	}
-
-	if fixedCount > 0 {
-		cacheDir := filepath.Join(filepath.Dir(azureYamlPath), ".azure", "cache")
-		cacheManager, cacheErr := cache.NewCacheManagerWithOptions(cache.CacheOptions{Enabled: true, CacheDir: cacheDir})
-		if cacheErr == nil {
-			if err := cacheManager.ClearCache(); err != nil && !cliout.IsJSON() {
-				cliout.Warning("Failed to clear cache: %v", err)
-			}
-		}
-	}
-
-	if !cliout.IsJSON() {
-		cliout.Newline()
-		cliout.Section(cliout.IconCheck, "Re-checking requirements...")
-	}
-
-	allResults := make([]ReqResult, 0, len(reqs))
-	allSatisfied := true
-	for _, prereq := range reqs {
-		result := runner.checker.Check(prereq)
-		allResults = append(allResults, result)
-		if !result.Satisfied {
-			allSatisfied = false
-		}
-	}
-
-	if cliout.IsJSON() {
-		return cliout.PrintJSON(map[string]any{
-			"success":      fixedCount > 0,
-			"fixed":        fixedCount,
-			"total":        len(failedReqs),
-			"allSatisfied": allSatisfied,
-			"fixes":        fixResults,
-			"results":      allResults,
-		})
-	}
-
-	cliout.Newline()
-	if fixedCount > 0 {
-		cliout.Success("Fixed %d of %d issues!", fixedCount, len(failedReqs))
-	} else {
-		cliout.Warning("Could not automatically fix any issues")
-	}
-
-	if !allSatisfied {
-		cliout.Newline()
-		cliout.Info("%s Next steps:", cliout.IconBulb)
-		cliout.Item("1. Run suggested install commands above")
-		cliout.Item("2. Restart your terminal to refresh PATH")
-		cliout.Item("3. Run 'azd app reqs' again to verify")
-		return fmt.Errorf("not all requirements satisfied")
-	}
-
-	cliout.Newline()
-	cliout.Success("All requirements now satisfied!")
-	cliout.Newline()
-	cliout.Info("ℹ️  Note: Tools may not be available in THIS terminal session")
-	if runtime.GOOS == osWindows {
-		cliout.Info("   To refresh PATH in your current PowerShell session, run:")
-		cliout.Info("   %s$env:PATH = [System.Environment]::GetEnvironmentVariable(\"Path\",\"Machine\") + \";\" + [System.Environment]::GetEnvironmentVariable(\"Path\",\"User\")%s", cliout.Dim, cliout.Reset)
-		cliout.Info("   Or simply restart your terminal")
-	} else {
-		cliout.Info("   To use the tools immediately, restart your terminal or source your shell profile")
-	}
-
-	return nil
 }
