+++ cli/src/cmd/app/commands/run.go
@@ -2,27 +2,17 @@ package commands
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"os"
-	"os/signal"
 	"path/filepath"
 	"sort"
 	"strings"
-	"sync"
-	"syscall"
-	"time"
 
-	"github.com/jongio/azd-app/cli/src/internal/dashboard"
 	"github.com/jongio/azd-app/cli/src/internal/detector"
-	"github.com/jongio/azd-app/cli/src/internal/executor"
-	"github.com/jongio/azd-app/cli/src/internal/notifications"
 	"github.com/jongio/azd-app/cli/src/internal/orchestrator"
 	"github.com/jongio/azd-app/cli/src/internal/service"
 	"github.com/jongio/azd-app/cli/src/internal/serviceinfo"
-	"github.com/jongio/azd-core/browser"
 	"github.com/jongio/azd-core/cliout"
-	"github.com/jongio/azd-core/registry"
 	"github.com/jongio/azd-core/yamlutil"
 
 	"github.com/spf13/cobra"
@@ -240,57 +230,6 @@ func detectServiceRuntimes(services map[string]service.Service, azureYamlDir, ru
 	return runtimes, nil
 }
 
-// executeAndMonitorServices starts services and monitors them until interrupted.
-func executeAndMonitorServices(ctx context.Context, runtimes []*service.ServiceRuntime, cwd string, azureYaml *service.AzureYaml, azureYamlDir string) error {
-	// Create logger
-	logger := service.NewServiceLogger(runVerbose)
-	logger.LogStartup(len(runtimes))
-
-	// Load environment variables
-	envVars, err := loadEnvironmentVariables()
-	if err != nil {
-		return err
-	}
-
-	// Orchestrate services with dependency ordering
-	result, err := service.OrchestrateServices(ctx, runtimes, azureYaml.Services, envVars, logger, runRestartContainers)
-	if err != nil {
-		return fmt.Errorf("service orchestration failed: %w", err)
-	}
-
-	// Validate that all services are ready
-	if err := service.ValidateOrchestration(result); err != nil {
-		service.StopAllServices(result.Processes)
-		return err
-	}
-
-	// Display service URLs (local + custom + Azure endpoints/domains)
-	serviceSummaries := buildServiceSummaries(cwd, azureYaml, result.Processes)
-	logger.LogSummary(serviceSummaries)
-
-	logger.LogReady()
-
-	// Execute postrun hook after all services are ready
-	if err := executePostrunHook(azureYaml, azureYamlDir); err != nil {
-		cliout.Warning("Postrun hook failed but services are running: %v", err)
-	}
-
-	// Display Functions/Logic Apps endpoints if any were discovered
-	if result.FunctionsParser != nil {
-		// Give functions a moment to finish startup logging
-		time.Sleep(2 * time.Second)
-
-		for name, process := range result.Processes {
-			if result.FunctionsParser.HasEndpoints(name) {
-				result.FunctionsParser.DisplayEndpoints(name, process.Port)
-			}
-		}
-	}
-
-	// Start dashboard and wait for shutdown
-	return monitorServicesUntilShutdown(result, cwd)
-}
-
 // loadEnvironmentVariables loads environment variables from --env-file if specified.
 func loadEnvironmentVariables() (map[string]string, error) {
 	if runEnvFile == "" {
@@ -438,339 +377,6 @@ func extractCustomConfig(azureYaml *service.AzureYaml) map[string]customEndpoint
 	return config
 }
 
-// monitorServicesUntilShutdown monitors all services with full process isolation.
-//
-// Process Isolation Design:
-//   - Each service runs in an independent goroutine with panic recovery
-//   - Service crashes/exits are logged but DON'T stop other services or the dashboard
-//   - Only user signals (Ctrl+C/SIGTERM) trigger coordinated shutdown of all services
-//   - Dashboard runs independently and survives individual service failures
-//
-// Lifecycle:
-//  1. Start monitoring goroutines (one per service + dashboard)
-//  2. Wait for user signal (Ctrl+C) or all services to naturally exit
-//  3. On signal: initiate graceful shutdown with 10-second timeout
-//  4. Stop all remaining services and dashboard
-//
-// This uses sync.WaitGroup (not errgroup) because we want all goroutines to complete
-// independently rather than failing fast on first error.
-func monitorServicesUntilShutdown(result *service.OrchestrationResult, cwd string) error {
-	// Create context that cancels on SIGINT/SIGTERM only
-	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
-	defer cancel()
-
-	var wg sync.WaitGroup
-	dashboardServer := dashboard.GetServer(cwd)
-
-	// Start notification manager for OS notifications on service issues
-	notifMgr, err := notifications.NewNotificationManager(
-		notifications.DefaultNotificationManagerConfig(cwd),
-	)
-	if err != nil {
-		cliout.Warning("Notifications unavailable: %v", err)
-	} else {
-		notifMgr.Start()
-		defer func() { _ = notifMgr.Stop() }()
-		// Notifications enabled silently - no need to announce
-	}
-
-	// Start dashboard monitoring (passes notifMgr to set URL after dashboard starts)
-	startDashboardMonitor(ctx, &wg, dashboardServer, notifMgr)
-
-	// Start service process monitors
-	startServiceMonitors(ctx, &wg, result.Processes, cwd)
-
-	// Wait for signal (context cancellation) or all services to complete
-	wg.Wait()
-
-	// Perform cleanup shutdown
-	return performGracefulShutdown(dashboardServer, result.Processes)
-}
-
-// startDashboardMonitor starts the dashboard server in a separate goroutine with panic recovery.
-func startDashboardMonitor(ctx context.Context, wg *sync.WaitGroup, dashboardServer *dashboard.Server, notifMgr *notifications.NotificationManager) {
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		defer func() {
-			if r := recover(); r != nil {
-				cliout.Error("Dashboard panic recovered: %v", r)
-			}
-		}()
-
-		dashboardURL, err := dashboardServer.Start()
-		if err != nil {
-			cliout.Warning("Dashboard unavailable: %v", err)
-			<-ctx.Done()
-			return
-		}
-
-		// Set dashboard URL for clickable notifications
-		if notifMgr != nil {
-			notifMgr.SetDashboardURL(dashboardURL)
-		}
-
-		cliout.Plain("  Dashboard  %s", dashboardURL)
-		cliout.Newline()
-
-		// Launch browser after dashboard is ready (if enabled)
-		browserLaunched := launchDashboardBrowser(dashboardURL)
-
-		// Show compact hints on a single line
-		if browserLaunched {
-			cliout.Hint("Press Ctrl+C to stop")
-		} else {
-			cliout.Hint("Press Ctrl+C to stop", "--web to open browser")
-		}
-
-		// Block until context is canceled
-		<-ctx.Done()
-	}()
-}
-
-// startServiceMonitors starts monitoring goroutines for all service processes.
-func startServiceMonitors(ctx context.Context, wg *sync.WaitGroup, processes map[string]*service.ServiceProcess, projectDir string) {
-	for name, process := range processes {
-		if process.Process == nil {
-			continue
-		}
-		wg.Add(1)
-		go monitorServiceProcess(ctx, wg, name, process, projectDir)
-	}
-}
-
-// performGracefulShutdown stops all services and dashboard with a timeout.
-// Returns nil due to process isolation design - individual failures are logged but don't fail the command.
-func performGracefulShutdown(dashboardServer *dashboard.Server, processes map[string]*service.ServiceProcess) error {
-	shutdownCtx, shutdownCancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer shutdownCancel()
-
-	cliout.Newline()
-	cliout.Newline()
-	cliout.Plain("Shutting down...")
-
-	// Stop dashboard
-	if stopErr := dashboardServer.Stop(); stopErr != nil {
-		cliout.Warning("Failed to stop dashboard: %v", stopErr)
-	}
-
-	// Stop all services with graceful timeout
-	if stopErr := shutdownAllServices(shutdownCtx, processes); stopErr != nil {
-		cliout.Warning("Some services failed to stop cleanly: %v", stopErr)
-	}
-
-	cliout.Success("All services stopped")
-	cliout.Newline()
-
-	// Clean up port assignments on clean shutdown
-	// Note: Port assignments are kept in the file for persistence across runs,
-	// but we don't release them here to allow quick restarts with same ports.
-	// Stale ports are cleaned up automatically after 7 days of inactivity.
-
-	// Always return nil due to process isolation design:
-	// Individual service crashes are logged but don't cause the run command to fail.
-	// Only return errors for infrastructure issues (dashboard, shutdown timeout, etc.)
-	return nil
-}
-
-// monitorServiceProcess monitors a single service process for exit or cancellation.
-// This function runs in its own goroutine with panic recovery to ensure one service
-// crash doesn't affect others (process isolation).
-func monitorServiceProcess(ctx context.Context, wg *sync.WaitGroup, serviceName string, proc *service.ServiceProcess, projectDir string) {
-	defer wg.Done()
-	defer func() {
-		if r := recover(); r != nil {
-			cliout.Error("Service monitor panic recovered for %s: %v", serviceName, r)
-		}
-	}()
-
-	// Wait for either process exit or context cancellation
-	// Use buffered channel to prevent goroutine leak
-	type exitResult struct {
-		exitCode int
-		err      error
-	}
-	waitDone := make(chan exitResult, 1)
-	go func() {
-		state, err := proc.Process.Wait()
-		if err != nil {
-			waitDone <- exitResult{exitCode: -1, err: fmt.Errorf("service %s exited with error: %w", serviceName, err)}
-			return
-		}
-		exitCode := state.ExitCode()
-		if !state.Success() {
-			waitDone <- exitResult{exitCode: exitCode, err: fmt.Errorf("service %s exited with code %d: %s", serviceName, exitCode, state.String())}
-			return
-		}
-		waitDone <- exitResult{exitCode: 0, err: nil}
-	}()
-
-	select {
-	case result := <-waitDone:
-		// Service exited - record exit info in registry
-		reg := registry.GetRegistry(projectDir)
-		endTime := time.Now()
-
-		// Always record exit code and end time for build/task mode tracking
-		if regErr := reg.UpdateExitInfo(serviceName, result.exitCode, endTime); regErr != nil {
-			cliout.Warning("Failed to update exit info for %s: %v", serviceName, regErr)
-		}
-
-		// Get service mode from registry to determine appropriate status
-		entry, _ := reg.GetService(serviceName)
-		mode := ""
-		if entry != nil {
-			mode = entry.Mode
-		}
-
-		if result.err != nil {
-			// Update registry to trigger OS notification via state monitor
-			// CRITICAL FIX: Implement retry logic for registry updates
-			maxRetries := 3
-			retryDelay := 100 * time.Millisecond
-			var regErr error
-			for i := 0; i < maxRetries; i++ {
-				regErr = reg.UpdateStatus(serviceName, "error")
-				if regErr == nil {
-					break
-				}
-				if i < maxRetries-1 {
-					time.Sleep(retryDelay)
-					retryDelay *= 2 // Exponential backoff
-				}
-			}
-
-			if regErr != nil {
-				cliout.Error("Failed to update registry for %s after %d retries: %v", serviceName, maxRetries, regErr)
-				// As fallback, try to send direct notification if notification manager is available
-				// This ensures users are informed even if registry update fails
-			}
-
-			// Show mode-appropriate error message
-			switch mode {
-			case service.ServiceModeBuild:
-				cliout.Error("Build failed: %s (exit code %d)", serviceName, result.exitCode)
-			case service.ServiceModeTask:
-				cliout.Error("Task failed: %s (exit code %d)", serviceName, result.exitCode)
-			default:
-				cliout.Error("⚠️  %v", result.err)
-				cliout.Warning("Service %s stopped. Other services continue running.", serviceName)
-				cliout.Info("Press Ctrl+C to stop all services")
-			}
-		} else {
-			// Update registry for clean exit
-			// Use mode-appropriate status
-			var status string
-			switch mode {
-			case service.ServiceModeBuild:
-				status = "built"
-				// Don't print message - build completion is expected, status visible in dashboard
-			case service.ServiceModeTask:
-				status = "completed"
-				// Don't print message - task completion is expected, status visible in dashboard
-			default:
-				status = "stopped"
-				cliout.Info("Service %s exited cleanly", serviceName)
-			}
-
-			// CRITICAL FIX: Implement retry logic for clean exit registry updates
-			maxRetries := 3
-			retryDelay := 100 * time.Millisecond
-			var regErr error
-			for i := 0; i < maxRetries; i++ {
-				regErr = reg.UpdateStatus(serviceName, status)
-				if regErr == nil {
-					break
-				}
-				if i < maxRetries-1 {
-					time.Sleep(retryDelay)
-					retryDelay *= 2 // Exponential backoff
-				}
-			}
-
-			if regErr != nil {
-				cliout.Warning("Failed to update registry for %s after %d retries: %v", serviceName, maxRetries, regErr)
-			}
-		}
-		// Intentionally don't cancel context - other services should continue
-	case <-ctx.Done():
-		// Context canceled by signal - proceed to graceful shutdown
-		return
-	}
-}
-
-// shutdownAllServices stops all services with graceful timeout.
-// Runs all shutdowns in parallel goroutines and waits for all to complete.
-// Returns aggregated errors from any services that failed to stop cleanly.
-func shutdownAllServices(ctx context.Context, processes map[string]*service.ServiceProcess) error {
-	var shutdownErrors []error
-	var mu sync.Mutex
-	var wg sync.WaitGroup
-
-	for name, process := range processes {
-		wg.Add(1)
-		go func(serviceName string, proc *service.ServiceProcess) {
-			defer wg.Done()
-
-			if proc.Process == nil {
-				return
-			}
-
-			// Determine timeout from context
-			deadline, ok := ctx.Deadline()
-			timeout := service.DefaultStopTimeout
-			if ok {
-				timeout = time.Until(deadline)
-				if timeout < time.Second {
-					timeout = time.Second
-				}
-			}
-
-			if err := service.StopServiceGraceful(proc, timeout); err != nil {
-				mu.Lock()
-				shutdownErrors = append(shutdownErrors, fmt.Errorf("%s: %w", serviceName, err))
-				mu.Unlock()
-			}
-		}(name, process)
-	}
-
-	wg.Wait()
-
-	if len(shutdownErrors) > 0 {
-		return fmt.Errorf("failed to stop %d service(s): %w", len(shutdownErrors), errors.Join(shutdownErrors...))
-	}
-	return nil
-}
-
-// runAspireMode runs Aspire AppHost directly using dotnet run.
-func runAspireMode(ctx context.Context, rootDir string) error {
-	// Find Aspire AppHost project
-	aspireProject, err := detector.FindAppHost(rootDir)
-	if err != nil {
-		return fmt.Errorf("failed to search for Aspire AppHost: %w", err)
-	}
-
-	if aspireProject == nil {
-		return fmt.Errorf("no Aspire AppHost found - --runtime aspire requires an AppHost.cs or Program.cs file in a .csproj project")
-	}
-
-	cliout.Plain("Running Aspire in native mode")
-	cliout.Item("Directory: %s", aspireProject.Dir)
-	cliout.Item("Project: %s", aspireProject.ProjectFile)
-	cliout.Newline()
-	cliout.Plain("Aspire dashboard will start automatically")
-	cliout.Newline()
-
-	// Use executor to run dotnet with proper environment inheritance
-	args := []string{"run", "--project", aspireProject.ProjectFile}
-
-	cliout.Hint("Press Ctrl+C to stop")
-	cliout.Newline()
-
-	// Run dotnet and let it handle everything (inherits all azd env vars)
-	return executor.StartCommand(ctx, "dotnet", args, aspireProject.Dir)
-}
-
 // showDryRun displays what would be executed without starting services.
 func showDryRun(runtimes []*service.ServiceRuntime) error {
 	cliout.Section("🔍", "Dry-run mode: Showing execution plan")
@@ -786,125 +392,6 @@ func showDryRun(runtimes []*service.ServiceRuntime) error {
 	}
 
 	return nil
-}
-
-// executePrerunHook executes the prerun hook if configured.
-func executePrerunHook(azureYaml *service.AzureYaml, workingDir string) error {
-	return executeHook(azureYaml, azureYaml.Hooks, azureYaml.Hooks.GetPrerun(), "prerun", workingDir)
-}
-
-// executePostrunHook executes the postrun hook if configured.
-func executePostrunHook(azureYaml *service.AzureYaml, workingDir string) error {
-	return executeHook(azureYaml, azureYaml.Hooks, azureYaml.Hooks.GetPostrun(), "postrun", workingDir)
-}
-
-// executeHook executes a lifecycle hook with the given name and configuration.
-// This is a common helper function to avoid duplication between prerun and postrun hooks.
-func executeHook(azureYaml *service.AzureYaml, hooks *service.Hooks, hook *service.Hook, hookName, workingDir string) error {
-	if hooks == nil || hook == nil {
-		return nil // No hook configured
-	}
-
-	// Convert service.Hook to executor.Hook
-	convertedHook := convertHook(hook)
-	config := executor.ResolveHookConfig(convertedHook)
-	if config == nil {
-		return nil
-	}
-
-	// Build environment variables for the hook
-	// Following azd pattern: pass project directory and any other context
-	hookEnvVars := buildHookEnvironmentVariables(azureYaml, workingDir)
-	config.Env = hookEnvVars
-
-	return executor.ExecuteHook(context.Background(), hookName, *config, workingDir)
-}
-
-// buildHookEnvironmentVariables builds environment variables to pass to hooks
-// Following the pattern from azure/azure-dev
-func buildHookEnvironmentVariables(azureYaml *service.AzureYaml, workingDir string) []string {
-	envVars := []string{
-		fmt.Sprintf("%s=%s", executor.EnvProjectDir, workingDir),
-		fmt.Sprintf("%s=%s", executor.EnvProjectName, azureYaml.Name),
-	}
-
-	// Add count of services for context
-	if azureYaml.Services != nil {
-		envVars = append(envVars, fmt.Sprintf("%s=%d", executor.EnvServiceCount, len(azureYaml.Services)))
-	}
-
-	return envVars
-}
-
-// convertHook converts service.Hook to executor.Hook to avoid circular imports.
-func convertHook(h *service.Hook) *executor.Hook {
-	if h == nil {
-		return nil
-	}
-	return executor.NewHook(
-		h.Run,
-		h.Shell,
-		h.ContinueOnError,
-		h.Interactive,
-		convertPlatformHook(h.Windows),
-		convertPlatformHook(h.Posix),
-	)
-}
-
-// convertPlatformHook converts service.PlatformHook to executor.PlatformHook.
-func convertPlatformHook(ph *service.PlatformHook) *executor.PlatformHook {
-	if ph == nil {
-		return nil
-	}
-	return executor.NewPlatformHook(
-		ph.Run,
-		ph.Shell,
-		ph.ContinueOnError,
-		ph.Interactive,
-	)
-}
-
-// resolveBrowserTarget determines which browser target to use.
-// Browser is OFF by default. Only opens if --web flag is specified.
-func resolveBrowserTarget(_ *service.AzureYaml) browser.Target {
-	if runWeb {
-		return browser.TargetSystem
-	}
-	return browser.TargetNone
-}
-
-// launchDashboardBrowser launches the dashboard in the configured browser.
-// Returns true if browser was launched, false if not (e.g., target is none).
-func launchDashboardBrowser(dashboardURL string) bool {
-	// Parse azure.yaml to get project config for browser preference
-	azureYamlPath, err := findAzureYaml()
-	var azureYaml *service.AzureYaml
-	if err == nil {
-		azureYaml, _ = service.ParseAzureYaml(azureYamlPath)
-	}
-
-	// Resolve browser target using priority system
-	target := resolveBrowserTarget(azureYaml)
-
-	// If target is none, don't launch
-	if target == browser.TargetNone {
-		return false
-	}
-
-	// Display launch message
-	targetName := browser.GetTargetDisplayName(target)
-	cliout.Plain("  Opening in %s...", targetName)
-
-	// Launch browser (non-blocking)
-	if err := browser.Launch(browser.LaunchOptions{
-		URL:     dashboardURL,
-		Target:  target,
-		Timeout: 5 * time.Second,
-	}); err != nil {
-		cliout.Warning("Could not open browser: %v", err)
-		cliout.Info("Dashboard available at: %s", dashboardURL)
-	}
-	return true
 }
 
 // REMOVED: initializeAzureLogBuffer - deprecated v1 polling/WebSocket implementation
