+++ cli/src/internal/rpc/logs.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"sort"
 	"strings"
 	"sync"
 
@@ -42,141 +41,6 @@ const (
 	defaultGetLogsTail = 500
 	maxGetLogsTail     = 10_000
 )
-
-// LogSource is the narrow read/subscribe slice of service.LogManager that
-// LogsHandler needs. Production wires it to a per-call LogManager
-// (matching how dashboard/server_handlers.go and dashboard/server_websocket.go
-// resolve the manager); tests inject an in-memory stub.
-//
-// Subscribe returns the buffered channel that *service.LogBuffer.Subscribe
-// produces (capacity 100, broadcast drop-newest with timeout). Unsubscribe
-// MUST close the channel; the implementation's existing semantics
-// (logbuffer.go) already do this.
-type LogSource interface {
-	// GetRecent returns the last `tail` entries for one service. The
-	// bool reports whether the service exists in the manager (false ->
-	// NotFound on the wire).
-	GetRecent(serviceName string, tail int) ([]service.LogEntry, bool)
-	// GetAll returns the merged tail across all services (limit per
-	// service applied internally, matching LogManager.GetAllLogs).
-	GetAll(tail int) []service.LogEntry
-	// ServiceNames lists the buffer keys currently registered. Used
-	// for the all-services subscribe path so the handler can fan in
-	// across every buffer without poking LogManager internals.
-	ServiceNames() []string
-	// Subscribe creates a per-service subscription channel. Returns
-	// false if the service does not exist (NotFound on the wire).
-	Subscribe(serviceName string) (chan service.LogEntry, bool)
-	// Unsubscribe releases the channel and frees server-side resources.
-	// MUST be safe to call with a channel from a now-removed service
-	// (idempotent / no panic on missing buffer).
-	Unsubscribe(serviceName string, ch chan service.LogEntry)
-}
-
-// ClassificationStore is the read/write slice of azure.yaml-stored log
-// classifications LogsHandler needs. Production wires it to dashboard's
-// loadAzureYaml/saveAzureYaml pair (closing over the package-level
-// classificationsMu so REST and Connect handlers share the same lock);
-// tests inject an in-memory stub.
-//
-// SaveClassifications receives the FULL replacement slice; the handler
-// performs read-modify-write under the same mutex the legacy REST
-// handler holds, so concurrent Add/Delete operations do not race.
-type ClassificationStore interface {
-	// LoadClassifications returns the current rule list from azure.yaml.
-	// An empty slice is valid (no rules). Error means the YAML could not
-	// be read or parsed; the handler maps that to Internal.
-	LoadClassifications() ([]service.LogClassification, error)
-	// SaveClassifications persists the full replacement slice, atomically
-	// from the dashboard side (fileutil.AtomicWriteFile).
-	SaveClassifications(classifications []service.LogClassification) error
-}
-
-// PreferenceStore is the read/write slice of the user-preferences blob.
-// Production wires it to azdconfig.ConfigClient.{Get,Set}PreferenceSection
-// keyed by "logs"; tests inject an in-memory stub.
-//
-// The blob is opaque JSON; the handler decodes/encodes via protojson with
-// DiscardUnknown=true so older or future-schema keys round-trip without
-// breaking GetPreferences. See the proto comment on Preferences for why
-// theme + ui.grid_auto_fit live here despite being absent from the Go
-// dashboard.UserPreferences struct.
-type PreferenceStore interface {
-	// LoadPreferences returns the raw JSON blob, or nil/empty if no
-	// preferences have been stored yet. nil and zero-length both signal
-	// "use defaults".
-	LoadPreferences() ([]byte, error)
-	// SavePreferences persists the raw JSON blob.
-	SavePreferences(data []byte) error
-}
-
-// LogsStore bundles the three narrow interfaces LogsHandler needs so the
-// Mount() conditional stays a single nil-check. Production wires it via
-// LogsStoreFuncs (below) closing over dashboard.Server methods; tests
-// satisfy it with in-memory stubs that implement all three slices.
-type LogsStore interface {
-	LogSource
-	ClassificationStore
-	PreferenceStore
-}
-
-// LogsStoreFuncs adapts plain function values to LogsStore. Lets the
-// dashboard wire its private loadAzureYaml/saveAzureYaml/getOrCreateConfigClient
-// helpers as method values without exporting them or smuggling the
-// classifications mutex through a wrapper struct.
-//
-// Field semantics mirror the LogSource/ClassificationStore/PreferenceStore
-// interface methods one-for-one. Every field MUST be set; nil function
-// values cause LogsHandler RPCs to panic at call time, which is the
-// fail-loud signal we want for a misconfigured wiring.
-type LogsStoreFuncs struct {
-	GetRecentFn           func(serviceName string, tail int) ([]service.LogEntry, bool)
-	GetAllFn              func(tail int) []service.LogEntry
-	ServiceNamesFn        func() []string
-	SubscribeFn           func(serviceName string) (chan service.LogEntry, bool)
-	UnsubscribeFn         func(serviceName string, ch chan service.LogEntry)
-	LoadClassificationsFn func() ([]service.LogClassification, error)
-	SaveClassificationsFn func(classifications []service.LogClassification) error
-	LoadPreferencesFn     func() ([]byte, error)
-	SavePreferencesFn     func(data []byte) error
-}
-
-// GetRecent implements LogSource.
-func (f LogsStoreFuncs) GetRecent(serviceName string, tail int) ([]service.LogEntry, bool) {
-	return f.GetRecentFn(serviceName, tail)
-}
-
-// GetAll implements LogSource.
-func (f LogsStoreFuncs) GetAll(tail int) []service.LogEntry { return f.GetAllFn(tail) }
-
-// ServiceNames implements LogSource.
-func (f LogsStoreFuncs) ServiceNames() []string { return f.ServiceNamesFn() }
-
-// Subscribe implements LogSource.
-func (f LogsStoreFuncs) Subscribe(serviceName string) (chan service.LogEntry, bool) {
-	return f.SubscribeFn(serviceName)
-}
-
-// Unsubscribe implements LogSource.
-func (f LogsStoreFuncs) Unsubscribe(serviceName string, ch chan service.LogEntry) {
-	f.UnsubscribeFn(serviceName, ch)
-}
-
-// LoadClassifications implements ClassificationStore.
-func (f LogsStoreFuncs) LoadClassifications() ([]service.LogClassification, error) {
-	return f.LoadClassificationsFn()
-}
-
-// SaveClassifications implements ClassificationStore.
-func (f LogsStoreFuncs) SaveClassifications(classifications []service.LogClassification) error {
-	return f.SaveClassificationsFn(classifications)
-}
-
-// LoadPreferences implements PreferenceStore.
-func (f LogsStoreFuncs) LoadPreferences() ([]byte, error) { return f.LoadPreferencesFn() }
-
-// SavePreferences implements PreferenceStore.
-func (f LogsStoreFuncs) SavePreferences(data []byte) error { return f.SavePreferencesFn(data) }
 
 // LogsHandler implements azdappv1connect.LogsServiceHandler.
 //
@@ -572,261 +436,4 @@ func (h *LogsHandler) SavePreferences(
 	}
 
 	return connect.NewResponse(&v1.SavePreferencesResponse{Preferences: prefs}), nil
-}
-
-// localLogRing is a bounded drop-oldest ring buffer with a coalescing
-// notify channel. Pumps push entries; the main loop drains on notify.
-// Concurrent push from N pump goroutines + concurrent drain from one
-// consumer is safe under the mutex.
-//
-// Notify channel semantics: capacity 1, non-blocking send. A single
-// notify wakes the consumer to drain everything, including any pushes
-// that landed AFTER the consumer started its current drain. The "drain
-// everything available" pattern means we never need more than one
-// outstanding notify regardless of producer rate.
-type localLogRing struct {
-	mu      sync.Mutex
-	buf     []*v1.LogEntry
-	cap     int
-	head    int // index of the oldest entry when ring is full
-	dropped int64
-	notify  chan struct{}
-}
-
-func newLocalLogRing(capacity int) *localLogRing {
-	if capacity <= 0 {
-		capacity = 1
-	}
-	return &localLogRing{
-		buf:    make([]*v1.LogEntry, 0, capacity),
-		cap:    capacity,
-		notify: make(chan struct{}, 1),
-	}
-}
-
-// push appends an entry, dropping the oldest one if the ring is full.
-// Always non-blocking. Coalesces notifications so a burst of pushes
-// produces at most one wakeup per drain cycle.
-func (r *localLogRing) push(e *v1.LogEntry) {
-	r.mu.Lock()
-	if len(r.buf) >= r.cap {
-		// Ring full: overwrite oldest via index, O(1).
-		r.buf[r.head] = e
-		r.head = (r.head + 1) % r.cap
-		r.dropped++
-	} else {
-		r.buf = append(r.buf, e)
-	}
-	r.mu.Unlock()
-
-	select {
-	case r.notify <- struct{}{}:
-	default:
-		// Already a pending wakeup; the consumer's next drain picks
-		// this entry up too. Coalescing.
-	}
-}
-
-// drain returns all entries currently in the ring in insertion order plus
-// the cumulative dropped count. Resets the ring; the caller compares
-// dropped against its previously-observed value to compute the delta.
-func (r *localLogRing) drain() ([]*v1.LogEntry, int64) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if len(r.buf) == 0 {
-		return nil, r.dropped
-	}
-	n := len(r.buf)
-	out := make([]*v1.LogEntry, n)
-	if r.head == 0 || n < r.cap {
-		// Not wrapped yet or head is at start - simple copy
-		copy(out, r.buf)
-	} else {
-		// Ring has wrapped: linearize from head
-		copy(out, r.buf[r.head:])
-		copy(out[n-r.head:], r.buf[:r.head])
-	}
-	r.buf = r.buf[:0]
-	r.head = 0
-	return out, r.dropped
-}
-
-// droppedCount exposes the cumulative drop counter for tests.
-//
-//nolint:unused // referenced from logs_test.go
-func (r *localLogRing) droppedCount() int64 {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	return r.dropped
-}
-
-// clampGetLogsTail mirrors handleGetLogs' tail-parameter handling:
-// zero/negative -> 500, above 10k -> 10k. Returns the adjusted value
-// and a flag indicating whether adjustment happened.
-func clampGetLogsTail(requested int32) (int, bool) {
-	if requested <= 0 {
-		// Zero is the proto default - the caller didn't ask for
-		// anything specific, so substituting the default is not a
-		// clamp. Negative values are nonsensical but we treat them as
-		// "use default" rather than erroring; flagging them as clamped
-		// would be technically true but useless to clients (they can't
-		// retry with a smaller positive value to "uncramp").
-		return defaultGetLogsTail, false
-	}
-	if requested > maxGetLogsTail {
-		return maxGetLogsTail, true
-	}
-	return int(requested), false
-}
-
-// fromProtoLogLevel maps the wire enum to the string ValidateClassificationLevel
-// expects. Unknown / unspecified collapses to "" so validation rejects it
-// rather than silently coercing to a default.
-func fromProtoLogLevel(l v1.LogLevel) string {
-	switch l {
-	case v1.LogLevel_LOG_LEVEL_INFO:
-		return "info"
-	case v1.LogLevel_LOG_LEVEL_WARN:
-		return "warning"
-	case v1.LogLevel_LOG_LEVEL_ERROR:
-		return "error"
-	case v1.LogLevel_LOG_LEVEL_DEBUG:
-		// Debug is a valid LogLevel for log entries but NOT a valid
-		// classification level (ValidateClassificationLevel only
-		// accepts info/warning/error). Return "" so AddClassification
-		// rejects it with InvalidArgument instead of silently saving
-		// an unparseable rule.
-		return ""
-	default:
-		return ""
-	}
-}
-
-// toProtoLogLevel maps service.LogLevel (int) to the wire enum.
-func toProtoLogLevel(l service.LogLevel) v1.LogLevel {
-	switch l {
-	case service.LogLevelInfo:
-		return v1.LogLevel_LOG_LEVEL_INFO
-	case service.LogLevelWarn:
-		return v1.LogLevel_LOG_LEVEL_WARN
-	case service.LogLevelError:
-		return v1.LogLevel_LOG_LEVEL_ERROR
-	case service.LogLevelDebug:
-		return v1.LogLevel_LOG_LEVEL_DEBUG
-	default:
-		return v1.LogLevel_LOG_LEVEL_UNSPECIFIED
-	}
-}
-
-// classificationLevelToProto maps the string-typed classification level
-// to the wire enum. Unknown strings collapse to UNSPECIFIED so a future
-// stored value doesn't masquerade as info.
-func classificationLevelToProto(level string) v1.LogLevel {
-	switch strings.ToLower(strings.TrimSpace(level)) {
-	case "info":
-		return v1.LogLevel_LOG_LEVEL_INFO
-	case "warning":
-		return v1.LogLevel_LOG_LEVEL_WARN
-	case "error":
-		return v1.LogLevel_LOG_LEVEL_ERROR
-	default:
-		return v1.LogLevel_LOG_LEVEL_UNSPECIFIED
-	}
-}
-
-// toProtoLogEntry converts a single service.LogEntry to its proto
-// representation. Maps IsStderr to LogStream STDOUT/STDERR; sets
-// LogSource based on service.LogEntry.Source ("local"/"azure"/empty);
-// includes timestamp, service, level, and message.
-//
-// service.LogEntry.Sequence and AzureMetadata are intentionally NOT
-// surfaced here: this RPC is for LOCAL logs only (StreamLocalLogs /
-// GetLogs scope), so Azure-only fields would be permanently zero on the
-// wire and confuse consumers. The Azure surface gets its own RPC.
-func toProtoLogEntry(e service.LogEntry) *v1.LogEntry {
-	stream := v1.LogStream_LOG_STREAM_STDOUT
-	if e.IsStderr {
-		stream = v1.LogStream_LOG_STREAM_STDERR
-	}
-	src := v1.LogSource_LOG_SOURCE_UNSPECIFIED
-	switch e.Source {
-	case service.LogSourceLocal, "":
-		src = v1.LogSource_LOG_SOURCE_LOCAL
-	case service.LogSourceAzure:
-		src = v1.LogSource_LOG_SOURCE_AZURE
-	}
-	return &v1.LogEntry{
-		Service:   e.Service,
-		Message:   e.Message,
-		Level:     toProtoLogLevel(e.Level),
-		Timestamp: timestamppb.New(e.Timestamp),
-		Stream:    stream,
-		Source:    src,
-	}
-}
-
-// toProtoLogEntries converts a slice of LogEntry. Sorted by timestamp
-// for determinism (LogManager.GetAllLogs already sorts; per-buffer
-// GetRecent is naturally ordered, but the merged path benefits from a
-// belt-and-braces sort here).
-func toProtoLogEntries(entries []service.LogEntry) []*v1.LogEntry {
-	out := make([]*v1.LogEntry, len(entries))
-	// Stable sort to preserve relative order of entries with identical
-	// timestamps (e.g. burst-logged events from the same service).
-	sort.SliceStable(entries, func(i, j int) bool {
-		return entries[i].Timestamp.Before(entries[j].Timestamp)
-	})
-	for i := range entries {
-		out[i] = toProtoLogEntry(entries[i])
-	}
-	return out
-}
-
-// toProtoClassification maps a single service.LogClassification.
-func toProtoClassification(c service.LogClassification) *v1.Classification {
-	return &v1.Classification{
-		Text:  c.Text,
-		Level: classificationLevelToProto(c.Level),
-	}
-}
-
-// toProtoClassifications maps a slice of service.LogClassification.
-func toProtoClassifications(in []service.LogClassification) []*v1.Classification {
-	out := make([]*v1.Classification, len(in))
-	for i := range in {
-		out[i] = toProtoClassification(in[i])
-	}
-	return out
-}
-
-// defaultProtoPreferences returns the default Preferences value, mirroring
-// dashboard.getDefaultPreferences (logs_config.go) exactly. Drift between
-// these two is a bug; the test harness asserts equality of the relevant
-// fields.
-func defaultProtoPreferences() *v1.Preferences {
-	return &v1.Preferences{
-		Version: "1.0",
-		// theme: empty string ("system") is the documented default in
-		// the dashboard hook; absence of a stored value falls back to
-		// the user-agent's preferred colour scheme. Leaving it empty
-		// here matches that behaviour without forcing a value the
-		// legacy code never wrote.
-		Theme: "",
-		Ui: &v1.UIPreferences{
-			GridColumns:      2,
-			GridAutoFit:      false,
-			ViewMode:         "grid",
-			SelectedServices: []string{},
-		},
-		Behavior: &v1.BehaviorPreferences{
-			AutoScroll:      true,
-			PauseOnScroll:   true,
-			TimestampFormat: "hh:mm:ss.sss",
-		},
-		Copy: &v1.CopyPreferences{
-			DefaultFormat:    "plaintext",
-			IncludeTimestamp: true,
-			IncludeService:   true,
-		},
-	}
 }
